Implement event subcommand
The subcommand allows a user to wait for events generated by
cgroup.event_control; it exits when a status crosses a threshold
specified by a user.

This subcommand allows specifying a relative value as a threshold,
for example, +1M which means that it sets a threshold as a current
value plus one mega bytes.