Support memory.pressure_level
It is similar to an event notification facility of
memory.usage_in_bytes and memory.memsw.usage_in_bytes;
it notifies memory pressure events to users through
cgroup.event_control. A user specifies one of 'low',
'medium' and 'critical' level to the kernel and the
kernel notifies to the listener when the memory pressure
exceeds the specified level.